server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
redis>=5.0.0
pyyaml>=6.0.0
cryptography>=41.0.0
requests>=2.31.0
